refactor(phase-1): extract mutation handlers into useEventMutations + useScheduleMutations
Moves ~550 lines of inline mutation logic out of WorksCalendar.tsx into two
focused hooks:

- useEventMutations: emitEventSave, checkEventConflicts, handleEventSave,
  handleEventMove, handleEventResize, handleEventGroupChange, handleEventDelete,
  handleInlineSave, handleInlineDelete

- useScheduleMutations: handleShiftStatusChange, handleCoverageAssign,
  handleEmployeeAction, handleAvailabilitySave, handleScheduleEditorSave

WorksCalendar.tsx shrinks from 3,117 → 2,598 lines. No behaviour changes.
